requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.34.0
botocore==1.34.0
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.6
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-environ==0.12.0
django-extensions==3.2.3
django-hosts==6.0
django-js-asset==3.1.2
django-recaptcha==4.1.0
django-storages==1.14.5
django-widget-tweaks==1.5.0
idna==3.10
jmespath==0.10.0
kombu==5.5.2
mysqlclient==2.0.3
numpy==2.2.4
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
s3transfer==0.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==1.26.20
vine==5.1.0
watchtower==1.0.6
wcwidth==0.2.13
XlsxWriter==3.2.2
django-admin-rangefilter==0.13.2
django-storages
django-celery-beat
